fix bug
* fix因为更改了秒数数据类型而造成后台服务不能转换成class的bug

+++ MES-MonitoringService/Model/MachineStatusLog.cs
@@ -33,7 +33,7 @@ namespace MES_MonitoringService.Model
         public bool IsStopFlag { get; set; }
 
         [BsonElement("UseTotalSeconds")]
-        public int UseTotalSeconds { get; set; }
+        public decimal UseTotalSeconds { get; set; }
 
         /*上传及更新的标识*/
         /*-------------------------------------------------*/
